fix: address PR #230 review feedback (round 4)
Sixteen unresolved review threads across correctness, performance,
documentation, and test organization:

- Evaluator: AnchorScope now saves/restores both last_match_end AND
  base_offset so `indirect` inside a `use` subroutine re-enters root
  rules with base_offset=0 (RFn1).
- Offset resolution: resolve_offset_with_base switches to checked
  arithmetic and maps overflow to InvalidOffset rather than saturating
  into a spurious BufferOverrun (RU0).
- MagicDatabase: drop the redundant rules Vec field; root_rules is the
  sole storage built via Arc::from(rules.into_boxed_slice()) to avoid
  Arc::from-slice clones (Vdy, VeD).
- Format: Conv::Char uses pad_non_numeric (space-only) so %03c follows
  POSIX instead of zero-padding; Conv::Char doc now says full 0x00-0xff
  range via Latin-1 (VeM, RFn6).
- Docs: MetaType::Offset added to architecture.md inline-dispatch list;
  evaluator.md dispatch wording separated from output formatting;
  evaluate_rules usage examples updated to pass &mut EvaluationContext;
  checklist bullet includes offset (RFnh, RFnl, RFnx, RFnz).
- Grammar: meta-type directives reject attached operators (`default&0xf`
  no longer silently drops the mask) with a regression test (RUs).
- Release workflow: dist-workspace.toml pin for actions/upload-artifact
  bumped from v7.0.0 to v7.0.1; release.yml regenerated via `dist
  generate` (RU7).
- Property test: timeout doc comment rewritten to reflect that the
  test constructs an env-less context, so MetaType::Indirect/Use take
  the silent no-op path rather than recursing (VeW).
- Name table: scrub_nested_names is now called with the actual parent
  level rather than parent.level + 1 (RFn-).
- Test layout: engine test helpers moved to
  src/evaluator/engine/tests/helpers/meta.rs (RFn3); grammar meta-type
  parser tests moved to src/parser/grammar/tests/meta_types.rs (RFn8).
