+++ dist-workspace.toml
@@ -53,4 +53,4 @@ publish-jobs = [ "homebrew" ]
 "actions/checkout"                = "v6.0.2"
 "actions/download-artifact"       = "3e5f45b2cfb9172054b4087a40e8e0b5a5461e7c"
 "actions/attest-build-provenance" = "a2bbfa25375fe432b6a289bc6b6cd05ecd0c4c32"
-"actions/upload-artifact"         = "v7.0.0"
+"actions/upload-artifact"         = "v7.0.1"
